requirements.txt
streamlit>=1.32.0
langchain>=0.2.5
langchain-community>=0.2.5
langchain-text-splitters>=0.2.5
langgraph>=0.0.40
langchain-openai>=0.1.8
google-generativeai>=0.8.0
google-api-python-client>=2.187.0
httplib2>=0.31.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
pymongo>=4.7.0
python-dotenv>=1.0.1
pypdf>=4.2.0
python-docx>=1.1.2
docx2txt>=0.8
unstructured>=0.12.6
numpy>=1.26.0
pandas>=2.2.0
plotly>=5.20.0
tenacity>=8.2.3
requests>=2.31.0
fpdf2>=2.7.8
cachetools>=5.3.3
typing_extensions>=4.12.0
orjson>=3.10.0
torch>=2.2.0
python-dateutil>=2.8.2

